Stores language settings in ini.
Store language specific stuff in the Parser struct. If the ini doesnt exist, save the default values. Load settings from the ini. If the settings for a language doesnt exist, use the default values. When notepad++ is closed it save the values. This way new parsers can be added without removing the ini.